Fix PLSQL
